+++ packages/protocol/src/protocol/ExchangeManager.ts
@@ -520,7 +520,7 @@ export class ExchangeManager implements Transport.Provider {
                 }
 
                 // Report peer loss to the session manager; this notifies all (relevant) sessions for the peer
-                await this.#sessions.handlePeerLoss(session.peerAddress, cause, createdAt);
+                await this.#sessions.handlePeerLoss(session.peerAddress, cause, createdAt, exchange);
             },
 
             retry: number => this.#sessions.retry.emit(session, number),